toolbox/requirements.txt
gym==0.12.1
lz4
Markdown
matplotlib
opencv-python
pytest==4.3.1
ray==0.8.0
tensorflow-probability
yapf==0.27
numpy==1.16.0
ray[debug]
gym[mujoco]
gym[box2d]
pandas
yattag
box2d-py==2.3.8
sklearn
seaborn
distro
gym[atari]
gputil
ray[tune]
ray[rllib]
